feat : api/game/games api 가격 데이터 소수점 2째 자리까지만 제공, 데이터 제공 날짜 수정

See merge request s08-bigdata-dist-sub2/S08P22C106!49

+++ back/src/main/java/bunsan/returnz/domain/game/dto/GameGamerStockDto.java
@@ -20,8 +20,8 @@ public class GameGamerStockDto {
 	private String companyCode;
 	private Integer totalCount; // 보유 수
 	private Integer totalAmount; // 총 가격
-	private Integer averagePrice; // 평균 단가
-	private Integer valuation; // 평가손익
+	private Double averagePrice; // 평균 단가
+	private Double valuation; // 평가손익
 	private Double profitRate; // 손익비율
 
 	// public GamerStock toEntity(GameGamerStockDto){
